Fix assertion failure in test lang:api
Running test lang:api with `valgrind` causes assertion failure. When 2rd
argument of `owiz_drop()` is very big (or negative) while `callstack.regs.sp`
is small, there might be an underflow in the subtraction in `owiz_drop()`,
and the `callstack.regs.sp` would be increased incorrectly. This fix checks
whether an underflow occurs.